+++ src/map/pet.h
@@ -37,7 +37,7 @@ struct pet_evolve_data {
 };
 
 struct s_pet_db {
-	short class_;
+	int class_;
 	char name[NAME_LENGTH],jname[NAME_LENGTH];
 	int itemID;
 	int EggID;
@@ -160,7 +160,7 @@ struct pet_interface {
 	int (*select_egg) (struct map_session_data *sd, int egg_index);
 	int (*catch_process1) (struct map_session_data *sd, int target_class);
 	int (*catch_process2) (struct map_session_data *sd, int target_id);
-	bool (*get_egg) (int account_id, short pet_class, int pet_id );
+	bool (*get_egg) (int account_id, int pet_class, int pet_id );
 	int (*unequipitem) (struct map_session_data *sd, struct pet_data *pd);
 	int (*food) (struct map_session_data *sd, struct pet_data *pd);
 	int (*ai_sub_hard_lootsearch) (struct block_list *bl, va_list ap);
